+++ test/types/index.ts
@@ -2922,3 +2922,16 @@ type a = keyof {}
 		? true
 		: false = true
 }
+
+// group empty prefix
+{
+	const app = new Elysia()
+		.group('', (app) => {
+			return app.get('/ok', () => 'Hello World')
+		})
+		.listen(3000)
+
+	type Routes = keyof (typeof app)['~Routes']
+
+	expectTypeOf<Routes>().toEqualTypeOf<'ok'>()
+}